Fasten GUI by adding updates support in the server
Adds a new client/server message: getUpdates().

Adds a new class, Update, that describes the changes in a maze since
the last call to getUpdates().

Adds a queue to queue up the updates.

Modify the Gui to update only the queue of updates, instead of
generating a new gridlayoyt for the maze every time the hero moves.

TODO: Trial stop working, must add it to the update support.