Reset update queue on each move
Resetting the update queue when calling the server.getUpdate() is
dumb, because clients that do not use the update feature, like the
textual client are forced to call getUpdates anyway to prevent the
queue to grow up infinitely.

It is much better to reset the update queue on every move.